@@ -3,8 +3,8 @@ package com.example.sokol.monitor.DateTimePicker;
 import android.app.Dialog;
 import android.app.TimePickerDialog;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.DialogFragment;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.DialogFragment;
 import android.text.format.DateFormat;
 
 import java.util.Calendar;
