@@ -11,8 +11,8 @@ import android.media.AudioAttributes;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v4.app.NotificationCompat;
-import android.support.v4.content.ContextCompat;
+import androidx.core.app.NotificationCompat;
+import androidx.core.content.ContextCompat;
 import android.widget.RemoteViews;
 
 import com.example.sokol.monitor.model.CatData;
@@ -127,7 +127,7 @@ public class NotificationProvider extends BroadcastReceiver {
         }
 
         NotificationCompat.Builder B = new NotificationCompat.Builder(context, channelID)
-                .setVisibility(Notification.VISIBILITY_PUBLIC)
+                .setVisibility(NotificationCompat.VISIBILITY_PUBLIC)
                 .setSmallIcon(R.drawable.ic_wb_incandescent_white_24dp)
                 .setOngoing(true)
                 .setCustomContentView(rv)
@@ -147,27 +147,6 @@ public class NotificationProvider extends BroadcastReceiver {
         NotificationManager.notify(NOTIFICATION_ID_REMOTE_CONTROL, notif);
     }
 
-    /**
-     * checks whether the remote control notification is there or wasn't yet pushed/was cancelled
-     * should not be used to determine whether or not to update the notification!
-     * It would break theme changes.
-     *
-     * @return true if notification controls are already among the active notifications.
-     */
-//    public static boolean isControlNotificationAlreadyThere(Context context){
-//        NotificationManager NotificationManager =
-//                (NotificationManager) context.getSystemService(Context.NOTIFICATION_SERVICE);
-//        if(NotificationManager == null) return false;
-//
-//        StatusBarNotification[] activeNotifs =  NotificationManager.getActiveNotifications();
-//        if(activeNotifs == null || activeNotifs.length == 0) return false;
-//
-//        for(StatusBarNotification sn : activeNotifs){
-//            if(sn.getId() == NOTIFICATION_ID_REMOTE_CONTROL) return true;
-//        }
-//
-//        return false;
-//    }
     public static void showNotificationIfEnabled(Context context, boolean withSound) {
         if (!SettingsFragment.getShowNotification(context)) return;
 
